+++ tailwind.config.js
@@ -2,6 +2,9 @@
 export default {
   content: ['./index.html', './src/**/*.jsx'],
   theme: {
+    screens:{
+      'web':'700px'
+    },
     extend: {},
   },
   plugins: [],
